Fix bug that caused app to crash
There was previously a bug that would cause the app to crash if we
inserted a new post at the same time that we received a new post.  This
fixes that bug.  It turns out that instead of setting an object at the
current position in an array, we actually want to remove the current
object at that position and insert a new one.